requirements.txt
certifi==2019.11.28
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
h5py==2.10.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==2.6.0
MarkupSafe==1.1.1
numpy==1.18.2
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
python-dotenv==0.12.0
pyzmq==19.0.0
connexion==2.4.0
sqlalchemy==1.3.16
flask-sqlalchemy==2.4.1
matplotlib==3.1.2
pillow==7.1.1
psycopg2-binary==2.8.4
pytest==4.0.2
pymysql==0.9.3
pyzmq==19.0.0
swagger-ui-bundle==0.0.6
scipy==1.4.1
scikit-learn==0.21.3
tensorflow==2.1.0
Werkzeug==1.0.0
mtcnn==0.1.0